requirements.txt
# Vision and Image Processing
opencv-python==4.8.1.78
numpy==1.24.3

# Robot Control
irobot-edu-sdk==1.0.0

# Configuration
pyyaml==6.0.1

# Jupyter (for notebooks)
jupyter>=1.0.0
ipython>=8.0.0

# Optional: for better visualization
matplotlib>=3.5.0
pillow>=9.0.0
